[+] Feat: Add GDRCopy ABI header and refactor related types in nv_gpu_gdrcopy.cpp
